requirements.txt
psycopg2-binary==2.9.8
pygame==2.5.2
python-decouple==3.8
